minor rework - non-working version

+++ include/pico_mqtt.h
@@ -1,6 +1,8 @@
 #ifndef PICO_W_MQTT_H
 #define PICO_W_MQTT_H
 
+#include <stdint.h>
+
 #define MQTT_SERVER         "192.168.61.111"
 
 #define MQTT_TOPIC_LEN      100
@@ -31,13 +33,12 @@
 
 typedef struct MQTT_CLIENT_DATA_T *MQTT_client_handle_t;
 
-
-int MQTT_open(MQTT_client_handle_t *handle);
-int MQTT_publish(MQTT_client_handle_t handle, const char *topic, const char *payload);
-int MQTT_subscribe(MQTT_client_handle_t handle, const char *topic);
-int MQTT_unsubscribe(MQTT_client_handle_t handle, const char *topic);
+uint8_t MQTT_open(MQTT_client_handle_t *handle);
+uint8_t MQTT_publish(MQTT_client_handle_t handle, const char *topic, const char *payload);
+uint8_t MQTT_subscribe(MQTT_client_handle_t handle, const char *topic);
+uint8_t MQTT_unsubscribe(MQTT_client_handle_t handle, const char *topic);
 void MQTT_close(MQTT_client_handle_t handle);
-int MQTT_poll(MQTT_client_handle_t handle);
-int MQTT_reconnect(MQTT_client_handle_t *handle);
+uint8_t MQTT_poll(MQTT_client_handle_t handle);
+uint8_t MQTT_reconnect(MQTT_client_handle_t *handle);
 
 #endif
